pyproject.toml
[project]
name = "fujidrop"
version = "0.1.0"
readme = "README.md"
description = "A frame.io replacement for the X100VI"
requires-python = ">=3.13"
dependencies = [
    "aioauth-client>=0.29.0",
    "aiohttp>=3.11.12",
    "boto3>=1.36.20",
    "google-auth-oauthlib>=1.2.1",
    "google-oauth2l>=1.0.2",
    "jinja2>=3.1.5",
    "peewee>=3.17.9",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.0",
    "sanic[ext]>=24.12.0",
]

[project.scripts]
serve = "backend.server:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
# autoflake (remove unused imports)
# isort (I)
lint.extend-select = ["I", "F401"]
# Be permissive of long lines
lint.ignore = ["E501"]
show-fixes = true
line-length = 100

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

